requirements.txt
scikit-learn~=0.23.2
scikit-optimize
scikit-image~=0.16.2
log4python
numpy~=1.19.0
Pillow~=8.0.0
xlsxwriter
Image
#Workbook==train.train
#selenium==3.141.0
opencv-python
cv
#cv2
#tabulate
rpy2~=3.3.6
sklearn~=0.0
torch~=1.5.0
tqdm~=4.50.2
torchvision~=0.6.0
torchsummary~=1.5.1
imageio~=2.6.1
matplotlib~=3.1.3
log4p~=2019.7.13.3
joblib~=0.17.0